feat: add a dumb DRM buffer cursor mode
It showed that dmabuf-based cursors are problematic when put on a plane. It is
likely that there is a format issue.

As an alternative to make it work and to have a comparision for future
development on the dmabuf pipeline add functionality to put the cursor texture
instead into a dumb DRM buffer.